feat(copy-groups): drag/drop reorder de slaves con @dnd-kit + recreación de tablas en prod
Hallazgo crítico (resuelto): las tablas copy_group_* figuraban como aplicadas
en supabase_migrations pero NO existían en producción. /dashboard/copy-groups
estaba devolviendo 500 hace tiempo (UI huérfana).

Fixes
- supabase/migrations/024 — cast `l.copy_multiplier::numeric` en recursive
  CTE (PG rechazaba el type mismatch numeric(10,4) vs numeric, abortando
  la migration y dejando las tablas a medio crear).
- Re-aplicada via `supabase db query --file`. 7 tablas creadas: copy_groups,
  copy_group_nodes, _links, _versions, _snapshots, _events, _experiments.

Feature
- migration 112: ADD COLUMN sort_index INT DEFAULT 0 en copy_group_nodes +
  índice (copy_group_id, role, sort_index). Aplicada.
- PATCH /api/copy-groups/[id]/nodes/reorder: bulk reorder con Zod validation,
  defense-in-depth (verifica IDs vs copy_group_id), audit a copy_group_events
  con event_type='nodes_reordered'.
- /api/copy-groups/[id]/graph: ahora ORDER BY role DESC, sort_index ASC.
- @dnd-kit/core + sortable + utilities instalados.
- SortableSlavesList.client.tsx: lista vertical sortable con GripVertical
  handle, PointerSensor (distance:6) + KeyboardSensor para a11y. Update
  optimista + rollback con toast.error en falla. aria-busy + aria-label.
- CopyGroupDetailWorkspace: añade sección "Orden de slaves" debajo del grafo
  SVG. Drag actualiza UI inmediato, PATCH en background, recarga al ok.

Verificación
- npm run lint: 0 errores
- npm run build: /api/copy-groups/[id]/nodes/reorder aparece en routes
- Tablas confirmadas con `SELECT FROM pg_tables WHERE tablename LIKE 'copy_group%'`

+++ supabase/migrations/024_copy_groups_capital_algorithm.sql
@@ -286,7 +286,7 @@ as $$
   with recursive tree as (
     select
       l.child_account_id as account_id,
-      l.copy_multiplier as multiplier_path,
+      l.copy_multiplier::numeric as multiplier_path,
       1 as depth,
       l.parent_account_id as parent_account_id
     from public.copy_group_links l
